# Reduce CPU overhead on `allocOutputs` when there is no allocation domain