Shape inference benchmark:

main:
```C++
-------------------------------------------------------------------------------------------
Benchmark                                                 Time             CPU   Iterations
-------------------------------------------------------------------------------------------
LayerNormBackward_ShapeInference                        129 us          129 us         5434
LayerNormForward_ShapeInference                        46.1 us         46.0 us        15050
LayerNormBackward_NoShapeInferenceCachedBaseline       79.8 us         79.6 us         8682
LayerNormForward_NoShapeInferenceCachedBaseline        37.7 us         37.7 us        18564
```

this:
```C++
-------------------------------------------------------------------------------------------
Benchmark                                                 Time             CPU   Iterations
-------------------------------------------------------------------------------------------
LayerNormBackward_ShapeInference                        120 us          120 us         5769
LayerNormForward_ShapeInference                        33.9 us         33.9 us        21020
LayerNormBackward_NoShapeInferenceCachedBaseline       70.1 us         70.0 us         9947
LayerNormForward_NoShapeInferenceCachedBaseline        25.5 us         25.5 us        27247
```

Python benchmark as posted in https://github.com/NVIDIA/Fuser/issues/749#issuecomment-1689896825:

main:
```markdown
[------------------  -----------------]
           |  nvfuser 0.0.19+git874e5c2
1 threads: ----------------------------
      2    |             27.3          
      4    |             37.6          
      8    |             57.5          
      16   |             95.2          
      32   |            191.9          
      64   |            369.3          
      128  |            726.5          

Times are in microseconds (us).
```
this
```markdown
[------------------  -----------------]
           |  nvfuser 0.0.19+git284989d
1 threads: ----------------------------
      2    |             23.1          
      4    |             33.8          
      8    |             53.4          
      16   |             90.5          
      32   |            190.3          
      64   |            368.9          
      128  |            733.7 

Times are in microseconds (us).
```