@naoyam 2023-08-23
> The change makes sense, but why does the perf get degraded in the largest size case?

I misunderstood the table with thread counts. It degraded with 128 threads not with some large fusions. Is the difference likely just a noise?