@zasdfgbnm 2023-08-23
> I misunderstood the table with thread counts. It degraded with 128 threads not with some large fusions. Is the difference likely just a noise?

I think your understanding is correct. It is a fusion with 128 inputs. And also, I modified this PR and the benchmark numbers after requested your review. So in case you are looking at the old numbers, yes, it is regressed due to the added `FUSER_PERF_SCOPE` in `polymorphicValueToBytes`. After I comment them away, the number for 128 inputs has dropped from ~760us -> ~730us, which is pretty close to the main branch, and I would consider the difference on numbers just as noise.